Rename LspStdio to Server, move out LSP modules
