+++ haskell-ide-engine.cabal
@@ -20,10 +20,10 @@ flag pedantic
 library
   hs-source-dirs:      src
   exposed-modules:     Haskell.Ide.Engine.Channel
-                       Haskell.Ide.Engine.LSP.CodeActions
-                       Haskell.Ide.Engine.LSP.Completions
+                       Haskell.Ide.Engine.CodeActions
+                       Haskell.Ide.Engine.Completions
                        Haskell.Ide.Engine.Plugin.Base
-                       Haskell.Ide.Engine.LSP.Reactor
+                       Haskell.Ide.Engine.Reactor
                        Haskell.Ide.Engine.Options
                        Haskell.Ide.Engine.Plugin.ApplyRefact
                        Haskell.Ide.Engine.Plugin.Brittany
@@ -44,7 +44,7 @@ library
                        Haskell.Ide.Engine.Support.Hoogle
                        Haskell.Ide.Engine.Support.Fuzzy
                        Haskell.Ide.Engine.Support.HieExtras
-                       Haskell.Ide.Engine.Transport.LspStdio
+                       Haskell.Ide.Engine.Server
                        Haskell.Ide.Engine.Types
   other-modules:       Paths_haskell_ide_engine
   build-depends:       Cabal >= 1.22
